refactor(API_KEY): use api key from config or env
