--- src/containers/MusicSearch/API_KEY.js
@@ -1,2 +1,0 @@
-const API_KEY = 'key';
-export default API_KEY;
